draft integration with pytest
