+++ .travis.yml
@@ -14,7 +14,7 @@ install:
 - sudo ifconfig dummy0:0 10.0.0.2 netmask 255.255.255.248
 - sudo ifconfig dummy0:1 10.0.0.3 netmask 255.255.255.248
 - ifconfig
-- pip install asyncio tox flake8 pylint cryptography==2.1.1
+- pip install asyncio tox flake8 pylint cryptography==2.1.1 pytest pytest-asyncio
 language: python
 script:
-- PYTHONPATH='.' python tests/main.py
+- pytest
